django/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
environ==1.0
mysqlclient==2.2.0
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
